I am changing the default behavior of error and the default class of IgBLAST. I want to parse the version of IgBLAST and make sure Airr is using the version of IgBlast
