+++ setup.cfg
@@ -45,6 +45,7 @@ install_requires =
 	requests
 	python-levenshtein
 	pyarrow
+	semantic_version
 
 [options.packages.find]
 where=src
